// fe/src/app/api/model/productResponse.ts
/**
 * Shop App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ImageDTO } from './imageDTO';


export interface ProductResponse { 
    id?: number;
    sku?: string;
    name?: string;
    description?: string;
    price?: number;
    importPrice?: number;
    categoryId?: number;
    categoryName?: string;
    images?: Array<ImageDTO>;
}